Freshen generalized type variables after turning lambda into named function (#971)
In `DesugarFuns`, lambdas are translated to named functions as follows:
```
fun (x) {
...
}
```
becomes 
```
{
  fun f(x) {...}
  f[arg1, ... ,argn]
}
```
Here, f is generalized to some type \forall a1, ... an. A, and the `arg1` to `argn` type arguments are inserted based on the lambda's original type.

However, there is an issue where we may generalize type variables that aren't present in the type environment, but nevertheless appear elsewhere outside of `f` as flexible variables. The current generalization approach mutates all generalized type variables directly, turning them into rigid ones. This means that any variable `a` appearing outside of `f` that gets generalized is turned rigid, too, and is therefore unbound, causing IR type errors.

This PR adds `generalise_without_mutation` to `Generlise`, which does not mutate/rigidify the generalized type variables. Instead, it uses the type variable renaming infrastructure to replace the generalized type variables with fresh, rigid variables inside the body of `f`.

Note that this is only used for generalization done by `DesugarFuns`, while `TypeSugar` still mutates type variables as before. We need to investigate further if we need to do the same refreshing there, too.

This fixes #871.

Co-authored-by: Daniel Hillerström <1827113+dhil@users.noreply.github.com>