Wrap `@load` in an error handler. (#1001)
This patch fixes a couple of bugs w.r.t. to error printing/handling, which would be visible in the REPL, e.g.

```
sig f : () -> T
fun f() { Foo }

links> @load "foo.links";
unknown directive : @load
```

The directive `@load` is uncertainly not unknown. What has happened
here is that the type constructor `T` is unbound, which in
`desugarEffects` would raise `NotFound`, which in turn would propagate
all the way out to the REPL, which catches `NotFound`, however, it
excepts the `NotFound` exception to raised if the given directive is
not found in the association list `directives`.

This patch fixes the exception raising points in `desugarEffects` to
raise `UnboundTyCon` instead, and wraps `@load` in `Errors.display`.