Default HTTP response when process throws `Aborted`. (#1012)
This patch fixes issue #1011. Patch #947 modified the response code
for error messages to be 500. Every client call aborts its server
process, and when a process aborts it throws the exception `Aborted`,
meaning it would inadvertently set the response code to 500. In the
long term we should probably reconsider the design of processes,
remote calls, etc.