Tag jsonised server side values (#1015)
This patch introduces tags on jsonised server side values in order to easily recognise the object structure during client side deserialisation. It also restores the functionality of the examples page. The changes made in this patch and the previous patch #1009 exposed several bugs in the client interaction with server side values.

Co-authored-by: Daniel Hillerström <1827113+dhil@users.noreply.github.com>
Co-authored-by: Daniel Hillerström <1827113+dhil@users.noreply.github.com>
Co-authored-by: Daniel Hillerström <1827113+dhil@users.noreply.github.com>
Co-authored-by: Daniel Hillerström <1827113+dhil@users.noreply.github.com>
Co-authored-by: Daniel Hillerström <daniel.hillerstrom@ed.ac.uk>