Coping with deeply nested literal JSON objects. (#1019)
Literal JSON objects must be passed to the client as string literals,
which are supplied to `JSON.parse` in order to avoid crashing the V8
parser for JavaScript literal JSON objects (it is a recursive descent
parser, which `JSON.parse` apparently isn't).

Technically, we need to do this trick for all literal objects that the
JavaScript compiler produces, including records, lists, and
variants. This patch only does it for top-level let bindings.

Partial fix for #1013.