[WIP] Fix #1037 (#1039)
Generalizes the behavior of childNodes to apply to an arbitrary list of XmlItems.
Generalizes the behavior of getTextContent so that it can run on either client or server.
Both functions are now implemented in Links, using new library primitives itemChildNodes and itemTextContent that take a single XmlItem as input. 
Removes the old childNodes, getTextContent, and getChildNodes from the library and jslib.