Heterogeneous queries and delateralization (#936)
This PR adds to Links set-based database queries, which can be freely composed with (standard) bag queries. No new syntax is provided (presently) for such queries; however, the new `dedup` function in the prelude, which removes duplicates from a list, is recognized by the database query implementation and used to produce `select distinct/union` statements rather than the usual `select/union all`.

The internal representation of queries is extended with `Prom` and `Dedup` constructors which embed a set query in a bag query and deduplicate a bag query to obtain a set query. The normalization code has been changed to handle these extra constructors,
and some of the changes (particularly those involving `For` comprehensions are rather complicated (perhaps too much). The most radical change is that `norm` now accepts queries containing `For` terms (whereas previously it expected all the input comprehensions to be expressed using `ConcatMap`). Getting this extension to work correctly with an incremental changed proved rather difficult and it's probably a good idea to rewrite the code that performs the normalization of `For` from scratch.

In addition to the `flat` and `nested` query policies for homogeneous queries, two new experimental policies are provided: `mixing` and `delat`. Both new policies know how to handle sets and bags, but they differ in the treatment of "lateral" variables. While normalized homogeneous comprehensions always have base tables as their generators, normalized heterogeneous comprehensions allow general subqueries in that position, which makes it possible for a generator to reference a variable introduced by a sister generator as in (pseudocode)
```
for (x <-- table1, y <- Q(x))
 ...

```
To express such a query in SQL directly, we have to use the SQL:1999 keyword `lateral`, which is what the `mixing` policy does. If a DBMS does not support `lateral` (or is particularly inefficient when handling lateral subqueries),  the `delat` policy employs a multi-step "delateralization" technique to produce an equivalent query without lateral references.

The two new policies employ a simplified SQL code generator (`query/SimpleSqlGen.ml`) due to my belief that the original code could be rewritten in a cleaner way knowing the exact shape of a normalized query. However, it is not as complete as the corresponding code in `query/sql.ml` (a quick look suggests that it doesn't support indices and ranges). At some point, the two SQL generators will have to be merged.