Add option to save the output of the JavaScript compiler.
This patch makes it possible to run Links in a "compile only mode"
which will store the output of `Irtojs.generate_program` in a
file. Note that the generated file may not be directly runnable
without linking the runtime system first. As of writing, the runtime
system must be linked manually, if required.