Fixes #1075 (#1082)
This is straightforward: the mixing normaliser (`mixingQuery.ml`) has an issue with `For` when its body is a `Concat`: the comprehension would normally be split along concatenation, but this does not happen. 

The issue is caused by the changes made to normalisation to allow it to consider normal forms where comprehension generators can be complex queries: in order for these changes to work, we must add a recursive call to `MixingQuery.reduce_for_body` (which until now was non-recursive).

The old normaliser doesn't show this bug, for reasons not entirely clear (but related to the simpler structure of homogeneous bag normal forms). However it's now advisable that the code responsible for normalising comprehensions should be refactored to make its rationale more understandable (without a clear rationale, every further change is at risk of creating new bugs).

Co-authored-by: James Cheney <jcheney@inf.ed.ac.uk>
Co-authored-by: Daniel Hillerström <1827113+dhil@users.noreply.github.com>