JavaScript runtime cleanup (#1110)
This is a minor follow-up patch on #1087. It fixes a couple of additional bugs discovered in the runtime. It also gets rid of all uses of spread syntax, as it is not supported by some of the tools we wish to feed Links applications to. For the same reason this patch also gets rid of all uses of `for (... in ...)` in favour of ordinary `for`-loops.