Options for linking JS runtime (#1113)
This patch is a follow-up on patch #1083. This patch adds two new (experimental) options to Links `-Ljs-runtime` and `-Xcustom-js-runtime`. The former option instructs the Links compiler to statically link the JavaScript runtime into a JavaScript compilation artefact. The latter option lets the user override the default JavaScript runtime (i.e. `jslib.js`) with a runtime of their own choosing.