Uniform paths (#1153)
This patch resolves #1132 by precomputing the various paths used by Links. The paths are computed during compilation of Links itself. The paths change depending on whether Links was compiled via `opam install` or `make`/`dune build`. In the former case, all the paths are computed relative to the OPAM root of the running switch, whereas in the latter case the paths are computed relative to the source tree that `make` was invoked in. Consequently, Links no longer has to guess or compute paths during startup (well this is a truth with modifications, database drivers are still discovered dynamically, and the default config file is still implicitly resolved if built from source as the testing infrastructure depends on being able to control the config file via the environment variable `LINKS_CONFIG`). In addition, this patch also deprecates the `LINKS_LIB` environment variable, i.e. setting this variable no longer affects the runtime of Links.

Co-authored-by: Daniel Hillerström <daniel.hillerstrom@ed.ac.uk>