New syntax for handlers (#1151)
This PR changes the syntax for handlers.
Instead of
```ocaml
handle (...) {
  case Op(params, resumption) -> ...
  case Return(x) -> x
}
```
we write
```ocaml
handle (...) {
  case <Op(params) => resumption> -> ...
  case x -> x
}
```
We can also write `case <Op(params) -> resumption> -> ...`. For now, the semantics for `->` and `=>` are the same. This is in prevision for a future step : we will not have `shallowhandler` anymore and we will indicate if a resumption is shallow with `->` and deep with `=>`.

Co-authored-by: RJ <juder.orbion@gmail.com>