Polymorphic operations (#1163)
This patch adds support for polymorphic operations.

For this we have to annotate the cases in the handler with polymorphic operation types.

Example
```links
sig h : (Comp(b, {A:forall a. (a) => a, Fail:forall a. a|c})) {A{_}, Fail{_}|c}~> Maybe(b)
fun h (m) {
    handle(m()) {
        case <A(x) => k> : (forall a.  (a) => a) -> k (x)
        case <Fail> : (forall a. () => a) -> Nothing
        case x -> Just(x)
    }
 }

sig f : () {A:forall a. (a) => a|_}-> Int
fun f () {
    if (do A(true)) do A(1) else
    do A(42)
}
```
This is done by having actual operation patterns and using the type annotation system associated with it.

Co-authored-by: Daniel Hillerström <daniel.hillerstrom@ed.ac.uk>
Co-authored-by: RJ <juder.orbion@gmail.com>
Co-authored-by: Robin Jourde <63565023+Orbion-J@users.noreply.github.com>