Remove links-mysql database driver (#1178)
Links currently has two database drivers for MySQL databases, provided by the packages links-mysql and links-mysql8. The former uses the opam package mysql, while the latter uses mysql8.

Unfortunately, the mysql opam package cannot be built on recent versions of Ubuntu anymore and has not been updated in a while. mysql8 is a fork of mysql that does work on more recent versions of Ubuntu, and supports MySQL version 8 (in addition to earlier versions).

Since mysql8 is backwards compatible with earlier versions of MySQL, its functionality subsumes that of mysql. However, since it's a fork of mysql we kept the links-mysql driver around in case the original mysql package gets updated eventually.

This PR is the result of giving up on this hope for the time being: It removes the links-mysql package and associated code within Links.

Since the CI is currently blocked by this, this PR also applies a temporary fix for #1179, by imposing a bound on the version of ppx_sexp_conv