Robust build
This patch makes the build infrastructure for Links more
robust. Previously, the build infrastructure would assume that it is
running in an environment populated by OPAM. This assumption is too
optimistic; for example a Docker container will give you a blank
environment by default. To make it easier to install Links in such
environments we now prefix dune build commands with `opam exec`. This
command resolves the installation directory of dune and makes sure it
is invoked with a suitable environment.

Furthermore, the `gen_linkspath.exe` install helper would only test
whether Links is being built by OPAM, and if that was not the case
then it would assume Links was being built from sources in a local
checkout of the git repository. This assumption is a tad optimistic
too. So now, if we cannot detect either an OPAM installation or the
git repository, then we fall back to a best effort guess based off the
current working directory.