Redirect command in MVU (#1193)
It's often the case that we want to change page as a result of, say, clicking a button in an MVU application. Currently we just use redirect directly as Links is impure, but this isn't a particularly idiomatic way of doing things and has led to a weird race condition in a student project.

This patch introduces a Redirect command, which can be accessed through MvuCommands.changePage(url). The redirect is then processed in the MVU event loop as would be expected.