User-definable abstract types
This patch makes it possible to declare abstract types in Links, e.g.

```links
typename MyAbstractType;
```

The primary motivation for adding them to be able to give types to
alien data objects.

The representation of user-definable abstract types is distinct from
the built-in abstract types. In a future patch we ought to reconcile
these two representations.