Fix CI failures and restore compatibility with OCaml 5.3 (#1212)
This patch bundles #1199, #1204, and #1208 into one.

## OCaml 5.3 compability

This patch makes it possible to compile Links with the recently released OCaml 5.3.0 compiler. In OCaml 5.3.0 the token `effect` has become a keyword, and thus cannot be used as the name of a binder anymore.

## CI workflow changes
Github is deprecating the CI image `ubuntu-20.04`:

> The Ubuntu 20.04 runner image will be fully unsupported by April 1, 2025. To raise awareness of the upcoming removal, we will temporarily fail jobs using Ubuntu 20.04. Builds that are scheduled to run during the brownout periods will fail. The brownouts are scheduled for the following dates and times:
>
>    March 4 14:00 UTC – 22:00 UTC
>    March 11 13:00 UTC – 21:00 UTC
>    March 18 13:00 UTC – 21:00 UTC
>    March 25 13:00 UTC – 21:00 UTC
>
> What you need to do
>
> Jobs using the ubuntu-20.04 YAML workflow label should be updated to ubuntu-22.04, ubuntu-24.04 or ubuntu-latest. You can always get up-to-date information on our tools by reading about the software in the runner images repository. Please contact GitHub Support if you run into any problems or need help. 

This patch updates the CI to use the `ubuntu-latest` image. 

In addition, this patch changes the OCaml compiler used for testing to 5.3.0 and upgrades the version of `ocaml/setup-ocaml` to version 3, which contains several fixes for the OCaml 5.3 toolchain. I have also disabled ocamlformat as the version we were using is incompatible with  OCaml 5.3 and adopting a new version required reformatting the lens files.