Fix higher-order CPS (#1217)
This patch fixes a bug in the JavaScript backend wrt. the name generation logic in the higher-order continuation module. Previously, name binding the continuation could inadvertently drop a prefix of the static continuation.

I have manually tested the few web examples that we have, and verified that the test case in #1097 works as expected now.

Resolves #1097